working on list_movies

+++ src/api/movies.py
@@ -115,7 +115,7 @@ def list_movies(
           dictionary["movie_id"] = int(movies["movie_id"])
           dictionary["movie_title"] = movies["title"]
           dictionary["year"] = movies["year"]
-          dictionary["imdb_rating"] = int(movies["imdb_rating"])
+          dictionary["imdb_rating"] = float(movies["imdb_rating"])
           dictionary["imdb_votes"] = int(movies["imdb_votes"])
           json.append(dictionary)
